requirements.txt
Django==4.2
djangorestframework==3.14.0
django-ckeditor==6.7.0
redis==5.0.1
django-redis==5.2.0
googletrans==4.0.0-rc1
pytest==7.4.0
flake8==6.1.0
